social_media_api/requirements.txt
Django>=3.0,<4.0
djangorestframework>=3.12,<4.0
djangorestframework-simplejwt>=4.0,<5.0
